+++ fj-doc-freemarker/src/main/java/org/fugerit/java/doc/freemarker/tool/FreeMarkerTemplateSyntaxVerifier.java
@@ -5,14 +5,24 @@ import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.Version;
 import lombok.extern.slf4j.Slf4j;
+import org.fugerit.java.core.cfg.ConfigRuntimeException;
 import org.fugerit.java.core.function.SafeFunction;
+import org.fugerit.java.core.io.FileIO;
+import org.fugerit.java.core.lang.helpers.BooleanUtils;
+import org.fugerit.java.core.lang.helpers.StringUtils;
 import org.fugerit.java.core.util.result.Result;
+import org.fugerit.java.doc.base.config.DocConfig;
+import org.fugerit.java.doc.base.process.DocProcessContext;
 import org.fugerit.java.doc.freemarker.config.FreeMarkerConfigStep;
+import org.fugerit.java.doc.freemarker.helper.FreeMarkerDocProcess;
 import org.fugerit.java.doc.freemarker.tool.verify.VerifyTemplateInfo;
 import org.fugerit.java.doc.freemarker.tool.verify.VerifyTemplateOutput;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.Properties;
 
 /**
@@ -43,6 +53,16 @@ public class FreeMarkerTemplateSyntaxVerifier {
      * It will check only templates with this pattern, (regex check on fileName, i.e. ".{0,}[.]ftl")
      */
     public static final String PARAM_TEMPLATE_FILE_PATTERN = "templateFilePattern";
+
+    /**
+     * If set to true a report will be generated (and property 'reportOutputFolder' will be olso required).
+     */
+    public static final String PARAM_GENERATE_REPORT = "generateReport";
+
+    /**
+     * Output folder for the generated report.
+     */
+    public static final String PARAM_REPORT_OUTPUT_FOLDER = "reportOutputFolder";
 
     public static VerifyTemplateOutput doCreateConfigurationAndVerify(File baseFolder) {
         return new FreeMarkerTemplateSyntaxVerifier().createConfigurationAndVerify( baseFolder, new Properties() );
@@ -92,8 +112,36 @@ public class FreeMarkerTemplateSyntaxVerifier {
             }
             // setup iteration with baseFolder as currentFolder
             this.iterateVerify( output, baseFolder, cfg, baseFolder );
+            Collections.sort( output.getInfos(), Comparator.comparing(VerifyTemplateInfo::getTemplateId) );
+            // generate report
+            this.generateReport( output, params );
             return output;
         } );
+    }
+
+    public void generateReport( VerifyTemplateOutput output, Properties params ) throws IOException {
+        boolean generateReport = BooleanUtils.isTrue( params.getProperty( PARAM_GENERATE_REPORT ) );
+        if ( generateReport ) {
+            String reportOutputFolder = params.getProperty( PARAM_REPORT_OUTPUT_FOLDER );
+            if (StringUtils.isEmpty( reportOutputFolder )) {
+                throw new ConfigRuntimeException( String.format( "Parameter '%s' is required", PARAM_REPORT_OUTPUT_FOLDER ) );
+            } else {
+                File reportOutputFolderFile = new File( reportOutputFolder );
+                if ( !reportOutputFolderFile.exists() ) {
+                    log.info( "{} does not exist, create directories : {}", reportOutputFolder, reportOutputFolderFile.mkdirs() );
+                }
+                String chainId = "freemarker-verify-syntax-report";
+                String type = DocConfig.TYPE_HTML;
+                File reportFile = new File( reportOutputFolderFile, chainId+"."+type );
+                try (ByteArrayOutputStream buffer = new ByteArrayOutputStream()) {
+                    FreeMarkerDocProcess.getInstance().fullProcess( chainId, DocProcessContext.newContext( "output", output ), type, buffer );
+                    FileIO.writeBytes( buffer.toByteArray(), reportFile );
+                } catch ( Exception e ) {
+                    throw new ConfigRuntimeException(
+                            String.format( "Fail to generate report in folder %s, error : %s", reportOutputFolder, e.getMessage() ), e );
+                }
+            }
+        }
     }
 
     /**