+++ fj-doc-freemarker/src/test/java/test/org/fugerit/java/doc/freemarker/tool/TestFreeMarkerTemplateSyntaxyVerifier.java
@@ -1,6 +1,7 @@
 package test.org.fugerit.java.doc.freemarker.tool;
 
 import lombok.extern.slf4j.Slf4j;
+import org.fugerit.java.core.cfg.ConfigRuntimeException;
 import org.fugerit.java.core.function.SafeFunction;
 import org.fugerit.java.core.util.result.Result;
 import org.fugerit.java.doc.freemarker.tool.FreeMarkerTemplateSyntaxVerifier;
@@ -12,9 +13,9 @@ import test.org.fugerit.java.BasicTest;
 
 import java.io.File;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Properties;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 
 @Slf4j
 public class TestFreeMarkerTemplateSyntaxyVerifier extends BasicTest {
@@ -62,6 +63,8 @@ public class TestFreeMarkerTemplateSyntaxyVerifier extends BasicTest {
 	public void verifyTemplateFilePattern() {
 		Properties params = new Properties();
 		params.setProperty( FreeMarkerTemplateSyntaxVerifier.PARAM_TEMPLATE_FILE_PATTERN, ".{0,}[.]ftl" );
+		params.setProperty( FreeMarkerTemplateSyntaxVerifier.PARAM_GENERATE_REPORT, "1" );
+		params.setProperty( FreeMarkerTemplateSyntaxVerifier.PARAM_REPORT_OUTPUT_FOLDER, "target/report-1" );
 		FreeMarkerTemplateSyntaxVerifier verifier = new FreeMarkerTemplateSyntaxVerifier();
 		Arrays.asList( "src/test/resources/fj_doc_test/template-fail",
 				"src/test/resources/fj_doc_test/template-macro" ).forEach(
@@ -69,6 +72,14 @@ public class TestFreeMarkerTemplateSyntaxyVerifier extends BasicTest {
 					VerifyTemplateOutput output = this.verifyWorker( basePath,
 							f -> verifier.createConfigurationAndVerify( f, params ) );
 					Assert.assertEquals(Result.RESULT_CODE_OK, output.getResultCode());
+				}
+		);
+		// check missing report folder
+		params.remove( FreeMarkerTemplateSyntaxVerifier.PARAM_REPORT_OUTPUT_FOLDER );
+		Arrays.asList( "src/test/resources/fj_doc_test/template-fail" ).forEach(
+				basePath -> {
+					Assert.assertThrows( ConfigRuntimeException.class, () -> this.verifyWorker( basePath,
+							f -> verifier.createConfigurationAndVerify( f, params ) ) );
 				}
 		);
 	}
@@ -81,6 +92,24 @@ public class TestFreeMarkerTemplateSyntaxyVerifier extends BasicTest {
 		Assert.assertThrows( NullPointerException.class, () -> new VerifyTemplateInfo( null, "2", null ) );
 		Assert.assertThrows( NullPointerException.class, () -> new VerifyTemplateInfo( null, null ) );
 		Assert.assertThrows( NullPointerException.class, () -> new VerifyTemplateInfo( null, null, null ) );
+	}
+
+	@Test
+	public void failReport() {
+		VerifyTemplateOutput output = new VerifyTemplateOutput() {
+			@Override
+			public List<VerifyTemplateInfo> getInfos() {
+				if ( Boolean.TRUE ) {
+					throw new ConfigRuntimeException( "Scenario error" );
+				}
+				return super.getInfos();
+			}
+		};
+		FreeMarkerTemplateSyntaxVerifier verifier = new FreeMarkerTemplateSyntaxVerifier();
+		Properties params = new Properties();
+		params.setProperty( FreeMarkerTemplateSyntaxVerifier.PARAM_GENERATE_REPORT, "1" );
+		params.setProperty( FreeMarkerTemplateSyntaxVerifier.PARAM_REPORT_OUTPUT_FOLDER, "target/report-fail" );
+		Assert.assertThrows( ConfigRuntimeException.class, () -> verifier.generateReport( output, params ) );
 	}
 
 }